Remove react components from deserialized document
