+++ source/index.js
@@ -2,3 +2,4 @@ export { default as deserializeDocument } from './deserialize'
 export * from './fetch'
 export * from './reducer'
 export * from './action'
+export * from './utils'
