+++ source/utils/index.js
@@ -1,0 +1,3 @@
+import { RichText } from 'prismic-reactjs'
+
+export const PrismicRichText = ({ children }) => RichText.render(children)
